feat: add bag play controller in evaluator
